Fix long lines in PDF User Guide
Pandoc converts all code blocks as verbatim in Latex by default.
This causes that long lines were not wrapped at the end of the
page. In order to support line breaking there is the need to
enable the listings module in Pandoc.

These are the changes made to achieve this:
    - template.tex : Format for listings blocks
    - create_pdf.sh : Enable --listings flag in pandoc command
    - 3.1.1-octl.md : Format changes for better alignement.

Change-Id: I172213dcacf48e66dc98662dc640986fd8e64232
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>